+++ tests/CSVArrayTest.php
@@ -128,7 +128,7 @@ final class CSVArrayTest extends TestCase
         );
     }
 
-    public function testToArrayFillsUpEmptyValuesWithEmptyStrings(): void
+    public function testToArrayDefaultsMissingColumnsToEmptyStrings(): void
     {
         self::assertSame(
             [
@@ -140,11 +140,22 @@ final class CSVArrayTest extends TestCase
                     'Spalte1' => 'Wert12',
                     'Spalte2' => CSVArray::DEFAULT_EMPTY_VALUE,
                 ],
+                3 => [
+                    'Spalte1' => '',
+                    'Spalte2' => '',
+                ],
+                4 => [
+                    'Spalte1' => 'Wert14',
+                    'Spalte2' => 'Wert24',
+                ],
             ],
             CSVArray::toArray(
                 "Spalte1;Spalte2\r\n"
                 . "Wert11;Wert21\r\n"
-                . "Wert12\r\n",
+                . "Wert12\r\n"
+                . ";\r\n"
+                . "\r\n"
+                . "Wert14;Wert24;Wert34",
             )
         );
     }